// src/app/models/progress.ts
export class Progress {
    user_id: Number;
    less_1_prog: Number;
    less_2_prog: Number;
    less_3_prog: Number;
    less_4_prog: Number;
    less_5_prog: Number;
    less_6_prog: Number;
    less_7_prog: Number;
    less_8_prog: Number;
    case_1_num: Number;
    case_2_num: Number;
    case_3_num: Number;
    case_4_num: Number;
    case_5_num: Number;
    case_6_num: Number;
    case_7_num: Number;
    case_8_num: Number;
} 